revert(web): drop the sectioned bell list, keeping the real one
The list the bell renders goes back to what it is: one flat stack of
rows with Clear all. The section headers, the split into Needs
attention and Updates, and the Clear updates label described a list
this product does not have, so the extracted list component and its
story go with them.

What the bell keeps is what the projection already earns: a waiting
request sorts first, carries the attention hue on its own row, and
withholds dismiss until it resolves.
