+++ clients/web/src/i18n/locales/zh-TW/home.json
@@ -9,8 +9,7 @@
     "viewSchedule": "檢視排程",
     "viewSkill": "檢視技能",
     "markAllAsRead": "全部標示為已讀",
-    "clearAll": "全部清除",
-    "clearUpdates": "清除更新"
+    "clearAll": "全部清除"
   },
   "category": {
     "security": "安全性",
@@ -19,7 +18,8 @@
     "background": "背景",
     "system": "系統",
     "guardianAction": "需要監護人操作",
-    "guardianQuestion": "監護人問題"
+    "guardianQuestion": "監護人問題",
+    "guardianRequest": "監護人請求"
   },
   "homeRecapRow": {
     "actionsTitle": "更新動作"
@@ -40,9 +40,7 @@
     "ariaLabelUnread": "通知 (未讀)",
     "heading": "通知",
     "loadFailed": "無法載入通知。",
-    "actionFailed": "無法開始該對話。請再試一次。",
-    "needsAttention": "需要關注",
-    "updates": "更新"
+    "actionFailed": "無法開始該對話。請再試一次。"
   },
   "notificationsBellDetail": {
     "back": "返回通知"
